+++ src/infra/accounts/mysql/tests/repos/test_mysql_account_repository.rs
@@ -61,6 +61,21 @@ async fn test_insert_and_locate_github_account(mysql_pool: MySqlPool) {
 
 #[test_group::group(database, mysql)]
 #[test_log::test(sqlx::test(migrations = "../../../../migrations/mysql"))]
+async fn test_insert_and_locate_multiple_github_account(mysql_pool: MySqlPool) {
+    let harness = MySqlAccountRepositoryHarness::new(mysql_pool);
+
+    run_transactional(&harness.catalog, |catalog: Catalog| async move {
+        kamu_accounts_repo_tests::test_insert_and_locate_multiple_github_account(&catalog).await;
+        Ok(())
+    })
+    .await
+    .unwrap();
+}
+
+/////////////////////////////////////////////////////////////////////////////////////////
+
+#[test_group::group(database, mysql)]
+#[test_log::test(sqlx::test(migrations = "../../../../migrations/mysql"))]
 async fn test_insert_and_locate_account_without_email(mysql_pool: MySqlPool) {
     let harness = MySqlAccountRepositoryHarness::new(mysql_pool);