+++ src/infra/accounts/postgres/src/repos/postgres_account_repository.rs
@@ -140,6 +140,62 @@ impl AccountRepository for PostgresAccountRepository {
         }
     }
 
+    async fn get_accounts_by_ids(
+        &self,
+        account_ids: Vec<AccountID>,
+    ) -> Result<Vec<Account>, GetAccountByIdError> {
+        let mut tr = self.transaction.lock().await;
+
+        let connection_mut = tr
+            .connection_mut()
+            .await
+            .map_err(GetAccountByIdError::Internal)?;
+
+        let accounts_search: Vec<_> = account_ids
+            .iter()
+            .map(std::string::ToString::to_string)
+            .collect();
+
+        let account_rows = sqlx::query!(
+            r#"
+            SELECT
+                id as "id: AccountID",
+                account_name,
+                email as "email?",
+                display_name,
+                account_type as "account_type: AccountType",
+                avatar_url,
+                registered_at,
+                is_admin,
+                provider,
+                provider_identity_key
+            FROM accounts
+            WHERE id = ANY($1)
+            "#,
+            &accounts_search
+        )
+        .fetch_all(connection_mut)
+        .await
+        .int_err()
+        .map_err(GetAccountByIdError::Internal)?;
+
+        Ok(account_rows
+            .into_iter()
+            .map(|account_row| Account {
+                id: account_row.id,
+                account_name: AccountName::new_unchecked(&account_row.account_name),
+                email: account_row.email,
+                display_name: account_row.display_name,
+                account_type: account_row.account_type,
+                avatar_url: account_row.avatar_url,
+                registered_at: account_row.registered_at,
+                is_admin: account_row.is_admin,
+                provider: account_row.provider,
+                provider_identity_key: account_row.provider_identity_key,
+            })
+            .collect())
+    }
+
     async fn get_account_by_name(
         &self,
         account_name: &AccountName,