+++ CHANGELOG.md
@@ -5,10 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- `InitiatorFilterInput` now accept `[AccountID]` instead of `AccountName`
+  `AccountFlowFilters` now filter by `DatasetId` instead of `DatasetName`.
 ### Added
 - Added a public image with [sqlx-cli](/images/sqlx-cli)
 - Added a [configuration](/images/persistence-storage) of running a `kamu` API server along with a database,
   for persistent storage of data
+- New `listFlowInitiators` api to fetch all initiators of flows
 
 ## [0.183.0] - 2024-05-22
 ### Added